// weapp/miniprogram/pages/merchant/_main_shared/api/table.ts
export interface TableResponse {
    capacity: number
    created_at: string
    current_reservation_id?: number              // 当前预定ID
    description?: string
    id: number
    merchant_id: number
    minimum_spend?: number
    qr_code_url?: string
    status: string
    table_no: string
    table_type: string
    tags?: TableTag[]
    updated_at: string
}

/** 桌台标签 */
export interface TableTag {
    id: number
    name: string
    color?: string
}